Move trim-whitespace to its own folder
