+++ components/object-property-list/trim-whitespace/trim-whitespace-directive.js
